Modificacioó del nom de les variables amb final.

+++ src/SCI/Condicionals/pt3NouCursV2.java
@@ -14,11 +14,11 @@ public class pt3NouCursV2 {
         int placesLliures = alumnes * aules * plantes - matricules;
         int placesSobrants = (alumnes * aules * plantes - matricules) * -1;
         double plantesFaltants = Math.ceil((double) placesSobrants / aules / alumnes);
-        int plantesLliures = (int) placesLliures / aules / alumnes;
+        int plantesLliures = placesLliures / aules / alumnes;
 
         if (alumnes > 0 && aules > 0 && plantes > 0 && matricules > 0){
             if (alumnes * aules * plantes > matricules){
-                System.out.println("A l'institut sobren " + (int) plantesLliures + " plantes");
+                System.out.println("A l'institut sobren " + plantesLliures + " plantes");
             }else {
                 System.out.println("Calen " + (int) plantesFaltants + " plantes per acollir tots els alumnes matriculats");
             }
